requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioop-lts==0.2.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
discord.py==2.5.2
distro==1.9.0
fastapi==0.115.12
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
markdown-it-py==3.0.0
mcp==1.9.0
mdurl==0.1.2
multidict==6.4.3
openai==1.79.0
packaging==25.0
pluggy==1.6.0
propcache==0.3.1
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.1.0
python-multipart==0.0.20
regex==2024.11.6
requests==2.32.3
rich==14.0.0
shellingham==1.5.4
sniffio==1.3.1
sse-starlette==2.3.5
starlette==0.46.2
tiktoken==0.9.0
tqdm==4.67.1
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
yarl==1.20.0
